+++ transforms/universal/doc_id/dpk_doc_id/spark/local.py
@@ -15,7 +15,7 @@ import sys
 
 from data_processing.utils import ParamsUtils
 from data_processing_spark.runtime.spark import SparkTransformLauncher
-from dpk_doc_id.spark.transform import (
+from dpk_doc_id.spark.runtime import (
     DocIDSparkTransformConfiguration,
     doc_column_name_cli_param,
     hash_column_name_cli_param,
